Split GUI context menu tests
